# src/swagger/url-shortener-api.yml
openapi: 3.0.0
info:
  title: URL Shortener API Specification
  version: 1.0
servers:
  - url: http://localhost:8022/
    variables: {}
security:
  - Bearer: []
paths:
  /api/v1:
    post:
      tags:
        - Create
      summary: Creates shortened version of the provided source URL and returning it.
      operationId: create
      requestBody:
        description: Source URL for shortening
        content:
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: URL address was successfully shortened and server responded with description object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceResponse'
        "400":
          description: User is trying to shorten an invalid input data that couldn't be recognized as URL address (or this URL belongs do domain that couldn't be shortened).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/bulk:
    post:
      tags:
        - Create
      summary: Creates shortened versions of the provided source URLs and returning it.
      operationId: bulkCreate
      requestBody:
        description: List of source URL addresses for shortening separated with newline '\n' character.
        content:
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: URL addresses were successfully shortened and server responded with description object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferencesResponse'
        "400":
          description: User is trying to shorten an invalid input data contains strings that couldn't be recognized as URL addresses (or some of them are belongs do domain that couldn't be shortened).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/{id}:
    get:
      tags:
        - Resolve
      summary: Resolve short URL to its source.
      operationId: resolve
      parameters:
        - name: id
          in: path
          description: Identifier that was given in the response to the request of 'URL shortening'.
          required: true
          content:
            '*/*':
              schema:
                type: string
      responses:
        "200":
          description: In a case of success redirect to the source URL will be performed.
        "400":
          description: User has invoked API with incorrect input data. For instance, wrong link format has been provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Provided id doesn't exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      title: ErrorResponse
      description: Describes error that has appeared during the invocation of API.
      properties:
        timestamp:
          type: string
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
      required:
        - timestamp
        - status
        - error
        - message
        - path
    ReferenceResponse:
      type: object
      title: ReferenceResponse
      description: Response that contains shortened URL address of source URL address.
      properties:
        shortUrl:
          type: string
      required:
        - shortUrl
    ReferencesResponse:
      type: object
      title: ReferencesResponse
      description: Response that contains shortened URL addresses of source URL addresses.
      properties:
        shortUrl:
          type: array
          items:
            type: string
      required:
        - shortUrl